split lattice and solved maze
