+++ maze_dataset/maze/__init__.py
@@ -1,0 +1,6 @@
+from maze_dataset.maze.lattice_maze import (
+    LatticeMaze,
+    ConnectionList,   
+)
+
+from maze_dataset.maze.solved_maze import SolvedMaze
